Split audio generation validation
